Small refactor after code review
