+++ source/Assets/Plugins/Didomi/Tests/SyncUserAfterInitTestsSuite.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 30c297c55fa5f477a8954d8aabaf060a
+guid: d1dc1c7b35f02499f92f36463cc6a7c0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
