+++ source/Assets/Plugins/Didomi/Tests/SyncUserBaseTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8bfefb7444b014a8d8d3a2d5797c7ac1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
